+++ src/app/core/models/summary-list-card-action/summary-list-card.type.ts
@@ -1,0 +1,12 @@
+export type SummaryListCardActionContent = {
+  key: string;
+  value: string;
+};
+
+export type SummaryListCardAction = {
+  id?: string;
+  title: string;
+  status?: 'pending' | 'existing';
+  showValue?: boolean;
+  content: SummaryListCardActionContent[];
+};
