+++ src/app/shared/models/result-wording-section/result-section.types.ts
@@ -1,0 +1,13 @@
+import { TemplateSubstitution } from '@openapi';
+import { PendingResultRow } from '@shared-types/result-code/result-code-row';
+
+export type UpdateExistingResultWordingPayload = {
+  resultId: string;
+  resultCode: string;
+  wordingFields: TemplateSubstitution[];
+};
+
+export type ResultSectionSubmitPayload = {
+  pendingToCreate: PendingResultRow[];
+  existingToUpdate: UpdateExistingResultWordingPayload[];
+};
